[test] use a standalone alternate loading test
previously due JavaClass caching the alternate loading
did not trigger a full class setup (which correctly fails
since class-loader only loads the one class) leading to
an illegal access error trying to setup TestHelper's
inner classes
