+++ test/org/jruby/test/TestHelper.java
@@ -122,13 +122,6 @@ public class TestHelper {
         }
     }
 
-
-    public static Class loadAlternateClass() throws ClassNotFoundException {
-        AlternateLoader loader = new AlternateLoader();
-        Class klass = loader.loadClass("org.jruby.test.TestHelper");
-        return klass;
-    }
-
     /**
      * Used by JVM bytecode compiler tests to run compiled code
      */
@@ -173,63 +166,6 @@ public class TestHelper {
             return cl;
         }
 
-    }
-    private static class AlternateLoader extends ClassLoader {
-
-        protected Class findModClass(String name) throws ClassNotFoundException {
-           byte[] classBytes = loadClassBytes(name);
-           replace(classBytes, "Original", "ABCDEFGH");
-           return defineClass(name, classBytes, 0, classBytes.length);
-        }
-        private void replace(byte[] classBytes, String find, String replaceWith) {
-            byte[] findBytes = find.getBytes();
-            byte[] replaceBytes = replaceWith.getBytes();
-            for (int i=0; i<classBytes.length; i++) {
-                boolean match = true;
-                for (int j=0; j<findBytes.length; j++) {
-                    if (classBytes[i+j] != findBytes[j]) {
-                        match = false;
-                        break;
-                    }
-                }
-                if (match) {
-                    for (int j=0; j<findBytes.length; j++)
-                        classBytes[i+j] = replaceBytes[j];
-                    return;
-                }
-            }
-        }
-        @Override
-        public Class loadClass(String name) throws ClassNotFoundException {
-            if (name.equals("org.jruby.test.TestHelper"))
-                return findModClass(name);
-            return super.loadClass(name);
-        }
-        private byte[] loadClassBytes(String name) throws ClassNotFoundException {
-            InputStream stream = null;
-            try {
-                String fileName = name.replaceAll("\\.", "/");
-                fileName += ".class";
-                byte[] buf = new byte[1024];
-                ByteArrayOutputStream bytes = new ByteArrayOutputStream();
-                int bytesRead = 0;
-                stream = getClass().getResourceAsStream("/" + fileName);
-                while ((bytesRead = stream.read(buf)) != -1) {
-                    bytes.write(buf, 0, bytesRead);
-                }
-                return bytes.toByteArray();
-            } catch (Exception e) {
-                e.printStackTrace();
-                throw new ClassNotFoundException(e.getMessage(),e);
-            } finally {
-                if (stream != null)
-                    try {
-                        stream.close();
-                    } catch (IOException e1) {
-                        e1.printStackTrace();
-                    }
-            }
-        }
     }
 
     private static class TestHelperException extends RuntimeException {
